# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Malaika-booking API",
        "description": "",
        "contact": {
            "email": "kirian@fruntera.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/passports": {
            "post": {
                "tags": [
                    "Passports"
                ],
                "summary": "Crea un nuevo pasaporte",
                "description": "Crea un nuevo pasaporte",
                "operationId": "createPassport",
                "requestBody": {
                    "description": "Create passport",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number_passport",
                                    "client_id"
                                ],
                                "properties": {
                                    "client_id": {
                                        "description": "Client ID",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "number_passport": {
                                        "description": "Passport number",
                                        "type": "string",
                                        "example": "4567845454WW"
                                    },
                                    "birth": {
                                        "description": "Passport owner bith date",
                                        "type": "string",
                                        "example": "1975/12/01"
                                    },
                                    "issue": {
                                        "description": "Passport issue",
                                        "type": "string",
                                        "example": "2021/12/01"
                                    },
                                    "exp": {
                                        "description": "Passport expiration date",
                                        "type": "string",
                                        "example": "2026/12/01"
                                    },
                                    "nationality": {
                                        "description": "Passport owner nacinality",
                                        "type": "string",
                                        "example": "Spanish"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Passport created successfully"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/passports/{id}": {
            "get": {
                "tags": [
                    "Passports"
                ],
                "summary": "Retorna un pasaporte por ID",
                "description": "Retorna un pasaporte",
                "operationId": "getPassportById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de pasaporte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Passport retrived successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PassportDetailResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Passports"
                ],
                "summary": "Actualiza los datos del pasaporte",
                "description": "Actualiza los datos del pasaporte",
                "operationId": "updatePassport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del pasaporte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update passport",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number_passport"
                                ],
                                "properties": {
                                    "number_passport": {
                                        "description": "Passport number",
                                        "type": "string",
                                        "example": "4567845454WW"
                                    },
                                    "birth": {
                                        "description": "Passport owner bith date",
                                        "type": "string",
                                        "example": "1975/12/01"
                                    },
                                    "issue": {
                                        "description": "Passport issue",
                                        "type": "string",
                                        "example": "2021/12/01"
                                    },
                                    "exp": {
                                        "description": "Passport expiration date",
                                        "type": "string",
                                        "example": "2026/12/01"
                                    },
                                    "nationality": {
                                        "description": "Passport owner nacinality",
                                        "type": "string",
                                        "example": "Spanish"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Passport updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PassportDetailResource": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "Passport owner name",
                        "type": "string",
                        "example": "Selene"
                    },
                    "surname": {
                        "description": "Passport owner surname",
                        "type": "string",
                        "example": "Selenita"
                    },
                    "number_passport": {
                        "description": "Passport number",
                        "type": "string",
                        "example": "4567845454WW"
                    },
                    "birth": {
                        "description": "Passport owner bith date",
                        "type": "string",
                        "example": "1975/12/01"
                    },
                    "issue": {
                        "description": "Passport issue",
                        "type": "string",
                        "example": "2021/12/01"
                    },
                    "exp": {
                        "description": "Passport expiration date",
                        "type": "string",
                        "example": "2026/12/01"
                    },
                    "nac": {
                        "description": "Passport owner nacinality",
                        "type": "string",
                        "example": "Spanish"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PassportResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}